templates: Rework templates to reduce use of jinja templates in Vue
To avoid switching between vue and jinja inside of Vue convert
chart-card, chart-card-header and filter-item to vue components and
include them in the main application using jinja.

Loosely following the principle that a component is either fully Vue or
fully jinja but not mixed other than includes and passing template
context to javascript vars.

Related issues:
https://github.com/ThreeSixtyGiving/insights-ng/issues/1
https://github.com/ThreeSixtyGiving/insights-ng/issues/2