js: data-display: Add a link to GrantNav
https://github.com/ThreeSixtyGiving/insights-ng/issues/16

Area filters rely on extra keyword fields being added to GrantNav (ie.
deploy https://github.com/ThreeSixtyGiving/grantnav/pull/815).

Using filters that are not supported results in a message instead of the
link. Currently this includes orgtype, which we intend to leave this
way, and awardDates, which we plan to add shortly.