js/templates: Remove map and rework choropleth widget
Remove map widget in favour of a second choropleth on the data page.
This gives a better overview of the distribution of grant locations by
geographical areas.

The choropleth widget has been refactored to be more generic to be able
to operate on both home and data pages. Some corner cases remain still
to be neatened up.

Related: https://github.com/ThreeSixtyGiving/insights-ng/issues/62